refactor: estrai domain/inspect_utils + domain/readiness da cli/
Sposta _helpers (schema, validazione, coerenza) in domain/inspect_utils.py.
Sposta readiness_ops in domain/readiness.py.
cli/inspect/_helpers.py diventa re-export + wrapper _payload_for_year.
cli/inspect/readiness_ops.py diventa re-export puro.
MCP e CLI consumer aggiornati a importare da domain/.
Azzera dipendenza MCP→cli/ per readiness e inspect_utils.

8 file, 0 logica modificata.
